Re-enabled filtering, sorting strings, added a remove string button
Re-enabled the ability to sort strings by length (ascending/descending).

Re-enabled the ability to filter strings by strings that contained the searched for phrase.

Both functions updated to fit with our stringsTableModel.

Added a remove string button that takes the string out of the table along with its address.

Made a temporary button for "DeepResearchAction".  Will be properly implemented at a later date.