Changed StringSniperPlugin Structure and added several features to component provider
String Sniper Plugin now imports ghidra.program.model.listing.Program so that the Program plugin can now pass information on the program that ghidra is analyzing.

Component provider has a rework to now include the ability to direct the user to the location in Ghidra where a string is that appears in the search table with its address.

When clicking an accordion panel inside of results tab, you now see: an updated confidence score out of 10 (subject to change), the temporary URL hyperlink of where the string appears in Sourcegraph, and text space for our LLM assessment.