Final update
Connection is now for the most part complete.  plugin can read result and matches.json as well as displaying the opensource code in the results tab.

Future updates can work on connecting main.py script to searchforstringsaction.java action performed section so that the entire process is seamless and user only has to click one button.